core/utils/ directory created and some files moved to here
